infrastructure/dgtocean/data-cluster/src/docker/rpc/requirements.txt
kafka-python
python-dotenv
certifi
requests
tenacity
